requirements.txt
beautifulsoup4
line.py @ git+https://github.com/seriaati/line.py
python-dotenv